+++ http_block_router.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"crypto/tls"
 	"fmt"
 	"net"
 	"net/http"
@@ -19,25 +20,34 @@ import (
 	"github.com/multiformats/go-multiaddr"
 )
 
-const httpBlockProviderTimeout = 5 * time.Second
-
-type httpBlockProvider struct {
+type httpBlockRouter struct {
 	endpoint   string
 	endpointMa multiaddr.Multiaddr
 	peerID     peer.ID
 	httpClient *http.Client
 }
 
-func newHTTPBlockProvider(endpoint string, p peer.ID, client *http.Client) (httpBlockProvider, error) {
+const httpBlockRouterTimeout = 5 * time.Second
+
+// newHTTPBlockRouter returns a router backed by a trustless HTTP gateway
+// (https://specs.ipfs.tech/http-gateways/trustless-gateway/) at the specified
+// endpoint. If gateway responds to HTTP 200 to HTTP HEAD request, the
+// FindProviders returns a provider record with predefined peerID and gateway
+// URL represented as multiaddr  with /tls/http suffic.
+func newHTTPBlockRouter(endpoint string, p peer.ID, client *http.Client) (httpBlockRouter, error) {
 	if client == nil {
-		client = defaultHTTPBlockProviderClient()
+		client = defaultHTTPBlockRouterClient(false)
 	}
+	if client.Timeout == 0 {
+		client.Timeout = httpBlockRouterTimeout
+	}
+
 	u, err := url.Parse(endpoint)
 	if err != nil {
-		return httpBlockProvider{}, fmt.Errorf("failed to parse endpoint %s: %w", endpoint, err)
+		return httpBlockRouter{}, fmt.Errorf("failed to parse endpoint %s: %w", endpoint, err)
 	}
 	if u.Scheme != "http" && u.Scheme != "https" {
-		return httpBlockProvider{}, fmt.Errorf("unsupported scheme %s, only http and https are supported", u.Scheme)
+		return httpBlockRouter{}, fmt.Errorf("unsupported scheme %s, only http and https are supported", u.Scheme)
 	}
 
 	h := u.Hostname()
@@ -54,7 +64,7 @@ func newHTTPBlockProvider(endpoint string, p peer.ID, client *http.Client) (http
 	var port int
 	if u.Port() != "" {
 		if p, err := strconv.Atoi(u.Port()); err != nil {
-			return httpBlockProvider{}, fmt.Errorf("invalid port %s: %w", u.Port(), err)
+			return httpBlockRouter{}, fmt.Errorf("invalid port %s: %w", u.Port(), err)
 		} else {
 			port = p
 		}
@@ -73,22 +83,22 @@ func newHTTPBlockProvider(endpoint string, p peer.ID, client *http.Client) (http
 		// allow unencrypted HTTP for local debugging
 		tlsComponent = ""
 	} else {
-		return httpBlockProvider{}, fmt.Errorf("failed to parse endpoint %s: only HTTPS providers are allowed (unencrypted HTTP can't be used in web browsers)", endpoint)
+		return httpBlockRouter{}, fmt.Errorf("failed to parse endpoint %s: only HTTPS providers are allowed (unencrypted HTTP can't be used in web browsers)", endpoint)
 
 	}
 
 	var httpPathComponent string
 	if escPath := u.EscapedPath(); escPath != "" && escPath != "/" {
-		return httpBlockProvider{}, fmt.Errorf("failed to parse endpoint %s: only URLs without path are supported", endpoint)
+		return httpBlockRouter{}, fmt.Errorf("failed to parse endpoint %s: only URLs without path are supported", endpoint)
 	}
 
 	endpointMaStr := fmt.Sprintf("/%s/%s/tcp/%d%s/http%s", hostComponent, h, port, tlsComponent, httpPathComponent)
 
 	ma, err := multiaddr.NewMultiaddr(endpointMaStr)
 	if err != nil {
-		return httpBlockProvider{}, fmt.Errorf("failed to parse endpoint %s: %w", endpoint, err)
+		return httpBlockRouter{}, fmt.Errorf("failed to parse endpoint %s: %w", endpoint, err)
 	}
-	return httpBlockProvider{
+	return httpBlockRouter{
 		endpoint:   endpoint,
 		endpointMa: ma,
 		peerID:     p,
@@ -96,18 +106,26 @@ func newHTTPBlockProvider(endpoint string, p peer.ID, client *http.Client) (http
 	}, nil
 }
 
-func defaultHTTPBlockProviderClient() *http.Client {
+func defaultHTTPBlockRouterClient(insecureSkipVerify bool) *http.Client {
+	transport := http.DefaultTransport
+	if insecureSkipVerify {
+		transport = &http.Transport{
+			TLSClientConfig: &tls.Config{
+				InsecureSkipVerify: true, // Disable TLS cert validation for tests
+			},
+		}
+	}
 	return &http.Client{
-		Timeout: httpBlockProviderTimeout, // timeout hanging HTTP HEAD sooner than boxo/routing/http/server.DefaultRoutingTimeout
+		Timeout: httpBlockRouterTimeout, // timeout hanging HTTP HEAD sooner than boxo/routing/http/server.DefaultRoutingTimeout
 		Transport: &drclient.ResponseBodyLimitedTransport{
-			RoundTripper: http.DefaultTransport,
+			RoundTripper: transport,
 			LimitBytes:   1 << 12, // max 4KiB -- should be plenty for HEAD response
 			UserAgent:    "someguy/" + buildVersion(),
 		},
 	}
 }
 
-func (h httpBlockProvider) FindProviders(ctx context.Context, c cid.Cid, limit int) (iter.ResultIter[types.Record], error) {
+func (h httpBlockRouter) FindProviders(ctx context.Context, c cid.Cid, limit int) (iter.ResultIter[types.Record], error) {
 	req, err := http.NewRequestWithContext(ctx, "HEAD", fmt.Sprintf("%s/ipfs/%s?format=raw", h.endpoint, c), nil)
 	if err != nil {
 		return nil, err
@@ -136,4 +154,4 @@ func (h httpBlockProvider) FindProviders(ctx context.Context, c cid.Cid, limit i
 	return iter.ToResultIter(iter.FromSlice([]types.Record{})), nil
 }
 
-var _ providersRouter = (*httpBlockProvider)(nil)
+var _ providersRouter = (*httpBlockRouter)(nil)